Object to do all the filtering and stuff
This filters out all the games that all players don't all
own. It bases it off game title for filtering at the moment
which should be fine.

"Players" are added by adding their game collections which are
lists of Game objects. These are basically plain ruby objects but
it might be nice to use a library or something that we can use to
make interfacing with these objects easier (i.e. don't need to
implement all the fields). One alternative is override method_missing
but that might be naughty. ¯\(°_o)/¯