Refactors the game import pipeline to a module
The AsyncronousProcessing module simply moves all the boilerplate
into a single location and takes care of all the boring bits
like pushing errors into proper channels and stuff.

You can write your processing plain and simple within #process, and if
an error or timeout happens it'll be shoved off to the right place.

Keep in mind if you want this to run cleanly, you'll need to provide
a logger. If the implementor of AsynchronousProcessing doesn't set a
logger it will just raise errors instead of just swallowing them.